+++ crates/ffi/jlrs_sys/build.rs
@@ -2,16 +2,14 @@ use std::env;
 
 use cfg_if::cfg_if;
 use find_julia::{JuliaDir, Version, enable_version_cfgs};
-use jlrs_compat::{
-    MAX_MINOR_VERSION, MIN_MINOR_VERSION, STABLE_MAJOR_VERSION, STABLE_MINOR_VERSION,
-};
+use jlrs_compat::{DOCS_MINOR_VERSION, MAJOR_VERSION, MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION};
 
 fn main() {
     // Enable julia_1_x configs
-    enable_version_cfgs(MIN_MINOR_VERSION, MAX_MINOR_VERSION);
+    enable_version_cfgs(MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION);
 
     if building_docs() {
-        let version = Version::new(STABLE_MAJOR_VERSION, STABLE_MINOR_VERSION, 0, false);
+        let version = Version::new(MAJOR_VERSION, DOCS_MINOR_VERSION, 0, false);
         version.emit_metadata_unchecked();
         return;
     }