+++ crates/jlrs/build.rs
@@ -1,15 +1,13 @@
 use std::env;
 
 use find_julia::{JuliaDir, Version, enable_version_cfgs};
-use jlrs_compat::{
-    MAX_MINOR_VERSION, MIN_MINOR_VERSION, STABLE_MAJOR_VERSION, STABLE_MINOR_VERSION,
-};
+use jlrs_compat::{DOCS_MINOR_VERSION, MAJOR_VERSION, MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION};
 
 fn main() {
-    enable_version_cfgs(MIN_MINOR_VERSION, MAX_MINOR_VERSION);
+    enable_version_cfgs(MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION);
 
     if building_docs() {
-        let version = Version::new(STABLE_MAJOR_VERSION, STABLE_MINOR_VERSION, 0, false);
+        let version = Version::new(MAJOR_VERSION, DOCS_MINOR_VERSION, 0, false);
         version.emit_metadata_unchecked();
         return;
     }