+++ crates/jlrs_macros/build.rs
@@ -1,24 +1,22 @@
-use std::env;
+use std::{env, path::PathBuf};
 
 use find_julia::{JuliaDir, enable_version_cfgs};
-use jlrs_compat::{
-    MAX_MINOR_VERSION, MIN_MINOR_VERSION, STABLE_MAJOR_VERSION, STABLE_MINOR_VERSION,
-};
+use jlrs_compat::{DOCS_MINOR_VERSION, MAJOR_VERSION, MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION};
 
 fn main() {
-    enable_version_cfgs(MIN_MINOR_VERSION, MAX_MINOR_VERSION);
+    enable_version_cfgs(MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION);
 
-    if building_docs() {
-        let version =
-            find_julia::Version::new(STABLE_MAJOR_VERSION, STABLE_MINOR_VERSION, 0, false);
-        version.emit_metadata_unchecked();
-        return;
-    }
+    let julia_dir = if building_docs() {
+        JuliaDir::docs(MAJOR_VERSION, DOCS_MINOR_VERSION)
+    } else {
+        JuliaDir::from_detected().expect("Julia not detected by jl-sys")
+    };
 
-    JuliaDir::from_detected()
-        .expect("Julia not detected by jl-sys")
-        .version()
-        .emit_metadata_unchecked();
+    let out_path = PathBuf::from(env::var("OUT_DIR").unwrap()).join("version.rs");
+    julia_dir
+        .write_version_file(&out_path, MIN_MINOR_VERSION, NIGHTLY_MINOR_VERSION)
+        .expect("Unable to write version file");
+    julia_dir.version().emit_metadata_unchecked();
 }
 
 fn building_docs() -> bool {